Implement setup and UI in hoomd_bevy crate.
