+++ hoomd-bevy/README.md
@@ -1,0 +1,6 @@
+# hoomd-bevy
+
+Interface [hoomd-rs] with [bevy].
+
+[hoomd-rs]: https://github.com/glotzerlab/hoomd-rs
+[bevy]: https://github.com/bevyengine/bevy
